# docs/swagger.yaml
schemes:
  - http
swagger: '2.0'
info:
  description: server API for Question Board Application.
  title: Question Board
  contact: {}
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: '1.0'
host: localhost:8080
basePath: /api
paths:
...
  /admin/subject:
    get:
      security:
        - ApiKey: []
      description: Route Path for Get List of Subject, for Administrator only.
      tags:
        - Subject
      summary: Get All Subject
      responses:
        '200':
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/response.MessageData'
              - type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/definitions/response.SubjectWithoutTeacher'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '417':
          description: Expectation Failed
          schema:
            $ref: '#/definitions/response.Error'
    post:
      security:
        - ApiKey: []
      description: Route Path for Insert New Subject, for Administrator only.
      consumes:
        - application/json
      produces:
        - application/json
      tags:
        - Subject
      summary: Create New Subject
      parameters:
        - description: send request subject code and subject name
          name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/request.SubjectRequest'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '417':
          description: Expectation Failed
          schema:
            $ref: '#/definitions/response.Error'
  /admin/subject/{id}:
    get:
      security:
        - ApiKey: []
      description: Route Path for Get Subject Details By ID.
      tags:
        - Subject
      summary: Get Subject By ID
      parameters:
        - type: integer
          description: subject id
          name: id
          in: path
          required: true
      responses:
        '200':
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/response.MessageData'
              - type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/definitions/response.Subject'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '417':
          description: Expectation Failed
          schema:
            $ref: '#/definitions/response.Error'
  /admin/subject/{id}/delete:
    delete:
      security:
        - ApiKey: []
      description: Route Path for Delete Subject, for Administrator only.
      consumes:
        - application/json
      produces:
        - application/json
      tags:
        - Subject
      summary: Delete Subject
      parameters:
        - type: integer
          description: subject id
          name: id
          in: path
          required: true
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '417':
          description: Expectation Failed
          schema:
            $ref: '#/definitions/response.Error'
  /admin/subject/{id}/update:
    put:
      security:
        - ApiKey: []
      description: Route Path for Update Subject, for Administrator only.
      consumes:
        - application/json
      produces:
        - application/json
      tags:
        - Subject
      summary: Update Subject
      parameters:
        - type: integer
          description: subject id
          name: id
          in: path
          required: true
        - description: send request subject code and subject name
          name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/request.SubjectRequest'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '417':
          description: Expectation Failed
          schema:
            $ref: '#/definitions/response.Error'
  /admin/tag:
    get:
      security:
        - ApiKey: []
      description: Route Path for Get List of Tag, for Administrator only.
      tags:
        - Tag
      summary: Get All Tag
      responses:
        '200':
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/response.MessageData'
              - type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/definitions/models.Tag'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '417':
          description: Expectation Failed
          schema:
            $ref: '#/definitions/response.Error'
    post:
      security:
        - ApiKey: []
      description: Route Path for Insert New Tag, for Administrator only.
      consumes:
        - application/json
      produces:
        - application/json
      tags:
        - Tag
      summary: Create New Tag
      parameters:
        - description: send request tag code and tag name
          name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/models.Tag'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '417':
          description: Expectation Failed
          schema:
            $ref: '#/definitions/response.Error'
  /admin/tag/{id}/delete:
    delete:
      security:
        - ApiKey: []
      description: Route Path for Delete Tag, for Administrator only.
      consumes:
        - application/json
      produces:
        - application/json
      tags:
        - Tag
      summary: Delete Tag
      parameters:
        - type: integer
          description: tag id
          name: id
          in: path
          required: true
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '417':
          description: Expectation Failed
          schema:
            $ref: '#/definitions/response.Error'
  /admin/tag/{id}/update:
    put:
      security:
        - ApiKey: []
      description: Route Path for Update Tag, for Administrator only.
      consumes:
        - application/json
      produces:
        - application/json
      tags:
        - Tag
      summary: Update Tag
      parameters:
        - type: integer
          description: tag id
          name: id
          in: path
          required: true
        - description: send request tag code and tag name
          name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/models.Tag'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '417':
          description: Expectation Failed
          schema:
            $ref: '#/definitions/response.Error'
...
  /student/subject:
    get:
      security:
        - ApiKey: []
      description: Route Path for Get List of Student Subject, for Student only.
      tags:
        - Subject
      summary: Get Student Subject
      parameters:
        - description: user data for fetch subject
          name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/models.User'
      responses:
        '200':
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/response.MessageData'
              - type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/definitions/response.SubjectWithTeacher'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '417':
          description: Expectation Failed
          schema:
            $ref: '#/definitions/response.Error'
...
  /teacher/subject:
    get:
      security:
        - ApiKey: []
      description: Route Path for Get List of Teacher Subject, for Teacher only.
      tags:
        - Subject
      summary: Get Teacher Subject
      parameters:
        - description: user data for fetch subject
          name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/models.User'
      responses:
        '200':
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/response.MessageData'
              - type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/definitions/response.SubjectWithStudent'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '417':
          description: Expectation Failed
          schema:
            $ref: '#/definitions/response.Error'
definitions:
...
  models.Level:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
  models.Major:
    type: object
    properties:
      code:
        type: string
        example: INF
      id:
        type: integer
        example: 1
      name:
        type: string
        example: Informatics
...
  models.Subject:
    type: object
    properties:
      code:
        type: string
      id:
        type: integer
      major:
        $ref: '#/definitions/models.Major'
      major_id:
        type: integer
      name:
        type: string
      student:
        type: array
        items:
          $ref: '#/definitions/models.User'
      teacher_class:
        type: array
        items:
          $ref: '#/definitions/models.TeacherSubject'
  models.Tag:
    type: object
    properties:
      id:
        type: integer
        example: 1
      name:
        type: string
        example: Year
      value:
        type: string
        example: '2019'
  models.TeacherSubject:
    type: object
    properties:
      class:
        type: string
      subject:
        $ref: '#/definitions/models.Subject'
      subject_id:
        type: integer
      teacher:
        $ref: '#/definitions/models.User'
      teacher_id:
        type: integer
...
  models.User:
    type: object
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      email:
        type: string
      id:
        type: integer
      level:
        $ref: '#/definitions/models.Level'
      level_id:
        type: integer
      major:
        $ref: '#/definitions/models.Major'
      major_id:
        type: integer
      name:
        type: string
      password:
        type: string
      status:
        type: integer
      subject:
        type: array
        items:
          $ref: '#/definitions/models.Subject'
      tags:
        type: array
        items:
          $ref: '#/definitions/models.Tag'
      teacher_subject:
        type: array
        items:
          $ref: '#/definitions/models.TeacherSubject'
      updatedAt:
        type: string
...
  request.SubjectRequest:
    type: object
    properties:
      code:
        type: string
        example: INF2213
      id:
        type: integer
        example: 1
      major:
        $ref: '#/definitions/models.Major'
      name:
        type: string
        example: Logic Informatic
      student:
        type: array
        items:
          $ref: '#/definitions/request.User'
      teacher_class:
        type: array
        items:
          $ref: '#/definitions/request.TeacherSubject'
  request.TeacherSubject:
    type: object
    properties:
      class:
        type: string
        example: A
      teacher:
        $ref: '#/definitions/request.User'
  request.User:
    type: object
    properties:
      email:
        type: string
      id:
        type: integer
      level:
        $ref: '#/definitions/models.Level'
      name:
        type: string
      status:
        type: integer
      tags:
        type: array
        items:
          $ref: '#/definitions/models.Tag'
...
  response.Error:
    type: object
    properties:
      error: {}
      message:
        type: string
  response.MessageData:
    type: object
    properties:
      data: {}
      message:
        type: string
  response.MessageOnly:
    type: object
    properties:
      message:
        type: string
...
  response.Student:
    type: object
    properties:
      email:
        type: string
      id:
        type: integer
      major:
        $ref: '#/definitions/models.Major'
      name:
        type: string
      tags:
        type: array
        items:
          $ref: '#/definitions/models.Tag'
  response.Subject:
    type: object
    properties:
      code:
        type: string
      id:
        type: integer
      major:
        $ref: '#/definitions/models.Major'
      name:
        type: string
      student:
        type: array
        items:
          $ref: '#/definitions/response.Student'
      teacher_class:
        type: array
        items:
          $ref: '#/definitions/response.TeacherClass'
...
  response.SubjectWithStudent:
    type: object
    properties:
      code:
        type: string
      id:
        type: integer
      major:
        $ref: '#/definitions/models.Major'
      name:
        type: string
      student:
        type: array
        items:
          $ref: '#/definitions/response.Student'
  response.SubjectWithTeacher:
    type: object
    properties:
      code:
        type: string
      id:
        type: integer
      major:
        $ref: '#/definitions/models.Major'
      name:
        type: string
      teacher_class:
        type: array
        items:
          $ref: '#/definitions/response.TeacherClass'
  response.SubjectWithoutTeacher:
    type: object
    properties:
      code:
        type: string
      id:
        type: integer
      major:
        $ref: '#/definitions/models.Major'
      name:
        type: string
  response.Teacher:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
  response.TeacherClass:
    type: object
    properties:
      class:
        type: string
      teacher:
        $ref: '#/definitions/response.Teacher'
...
securityDefinitions:
  ApiKey:
    type: apiKey
    name: Authorization
    in: header